+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 #### Added
 - Temples icons
+- Amount on staking account row 
 
 ## [0.1.0-rc.1] - 2021-09-25
 #### Added